Create T-Phish
README.md
T-Phish
About T-Phish :
T-Phish is agreat and very usfull phishing tool for Mobile User. This is a modifyed Tool

Installation and usage guide
$ apt-get update -y

$ apt-get upgrade -y

$ pkg install git -y

$ git clone https://github.com/Stephin-Franklin/T-Phish

$ ls

$ cd T-Phish

$ ls

$ unzip T-Phish

$ cd T-Phish

$ unzip T-Phish.zip

$ chmod 777 start.sh

$ ./start.sh

$ ./phish.sh

Warning
This tool is only for educational purpose. If you use this tool for other purposes except education we will not be responsible in such cases.